Updated i18next definitions to 2.3.4.
