+++ i18next-browser-languagedetector/i18next-browser-languagedetector.d.ts
@@ -4,7 +4,7 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
 ///<reference path="../express/express.d.ts"/>
-///<reference path="../i18next/i18next.d.ts"/>
+///<reference path="../i18next/i18next-2.0.17.d.ts"/>
 
 declare namespace I18next {
     interface I18nextStatic extends i18nextBrowserLanguageDetector.I18nextStatic { }
