+++ i18next-sprintf-postprocessor/i18next-sprintf-postprocessor.d.ts
@@ -4,7 +4,7 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
 ///<reference path="../express/express.d.ts"/>
-///<reference path="../i18next/i18next.d.ts"/>
+///<reference path="../i18next/i18next-2.0.17.d.ts"/>
 
 declare namespace I18next {
 	interface I18nextOptions extends i18nextSprintfPostProcessor.I18nextOptions {}
